@@ -1,7 +1,7 @@
 {
-  "name": "flowbite-svelte-starter",
-  "version": "0.2.0",
-  "description": "SVG icon collection for Svelte",
+  "name": "svelte-icon-webkit",
+  "version": "0.1.0",
+  "description": "A collection of reusable Svelte components for building icon-based user interfaces in web applications.",
   "author": {
     "name": "Shinichi Okada",
     "email": "connect@codewithshin.com",
@@ -25,45 +25,45 @@
     "test:unit": "vitest"
   },
   "devDependencies": {
+    "@changesets/cli": "^2.27.1",
     "@playwright/test": "^1.42.1",
-    "@sveltejs/adapter-auto": "^3.1.1",
+    "@sveltejs/adapter-auto": "^3.2.0",
     "@sveltejs/kit": "^2.5.4",
     "@sveltejs/package": "^2.3.0",
     "@sveltejs/vite-plugin-svelte": "^3.0.2",
+    "@svitejs/changesets-changelog-github-compact": "^1.1.0",
+    "@types/eslint": "^8.56.6",
     "@typescript-eslint/eslint-plugin": "^6.21.0",
     "@typescript-eslint/parser": "^6.21.0",
     "autoprefixer": "^10.4.19",
     "eslint": "^8.57.0",
     "eslint-config-prettier": "^9.1.0",
-    "eslint-plugin-svelte": "^2.35.1",
-    "flowbite-svelte": "^0.44.24",
-    "flowbite-svelte-icons": "^1.5.0",
+    "eslint-plugin-svelte": "2.36.0-next.9",
     "postcss": "^8.4.38",
     "postcss-load-config": "^5.0.3",
     "prettier": "^3.2.5",
     "prettier-plugin-svelte": "^3.2.2",
-    "svelte": "^4.2.12",
+    "prettier-plugin-tailwindcss": "^0.5.13",
+    "publint": "^0.2.7",
+    "svelte": "5.0.0-next.85",
+    "svelte-5-ui-lib": "^0.3.7",
     "svelte-check": "^3.6.8",
-    "svelte-meta-tags": "^3.1.1",
+    "svelte-lib-helpers": "^0.4.5",
     "svelte-preprocess": "^5.1.3",
+    "svelte-rune-highlight": "^0.4.0",
     "tailwind-merge": "^2.2.2",
-    "tailwindcss": "^3.4.1",
+    "tailwindcss": "^3.4.3",
     "tslib": "^2.6.2",
     "typescript": "^5.4.3",
     "vite": "^5.2.6",
     "vitest": "^1.4.0"
   },
   "type": "module",
-  "dependencies": {
-    "@popperjs/core": "2.11.6",
-    "classnames": "^2.5.1",
-    "flowbite": "^1.8.1"
-  },
   "keywords": [
     "svelte",
     "sveltekit",
     "flowbite",
-    "flowbite-svelte",
+    "svelte-5-ui-lib",
     "tailwindcss"
   ],
   "repository": {
