@@ -1,6 +1,12 @@
 import { sveltekit } from '@sveltejs/kit/vite';
 import { defineConfig } from 'vite';
+import pkg from './package.json' assert { type: 'json' };
 
 export default defineConfig({
-  plugins: [sveltekit()]
+  plugins: [sveltekit()],
+  define: {
+		__NAME__: `"${pkg.name}"`,
+		__VERSION__: `"${pkg.version}"`,
+		__GITHUBURL__: `"${pkg.repository.url}"`
+	},
 });
